requirements 2.txt
# requirements.txt
streamlit>=1.38
google-cloud-vision>=3.7
python-dotenv>=1.0
pillow>=10.4
regex>=2024.7.24
google-genai>=0.5.0
natsort>=8.0
